🚀 MCP Server and EKS Tools
working mcp server and base k8s tools
